Added Util functions & Removed FinancialYear class
