+++ src/EPR.Calculator.API/Controllers/CalculatorInternalController.cs
@@ -69,7 +69,7 @@ namespace EPR.Calculator.API.Controllers
                     var stagingOrganisationData = this.wrapper.GetOrganisationData();
                     var calcOrganisationMaster = new CalculatorRunOrganisationDataMaster
                     {
-                        CalendarYear = FinancialYear.ToCalendarYearAsString("2024-25"), //Take the financial year from Calc Run table and Derive the Calendar year
+                        CalendarYear = Util.GetCalendarYear("2024-25"), //Take the financial year from Calc Run table and Derive the Calendar year
                         CreatedAt = DateTime.Now,
                         CreatedBy = request.UpdatedBy,
                         EffectiveFrom = DateTime.Now,
